Update to fetch correct data
Previously, the app was not fetching info on userName because it could not access the data. As the requests had to include the username, using user id, as the previous code, was unsuccessful. Had to store the userName from user in order for it to work. On the way, renamed the variable with a more precise name (by recomendation of my tutor)